sql + keycloack

+++ server/src/main/resources/db/V3__create_categoria.sql
@@ -1,5 +1,5 @@
 CREATE TABLE rh.categoria (
-	id numeric NOT NULL,
+	id int8 NOT NULL,
 	descricao varchar(255) NOT NULL,
 	CONSTRAINT categoria_pk PRIMARY KEY (id)
 );
