+++ server/src/main/resources/db/V4__create_qualificacao.sql
@@ -1,8 +1,8 @@
 CREATE TABLE rh.qualificacao (
-	id numeric NOT NULL,
+	id int8 NOT NULL,
 	descricao varchar(255) NOT NULL,
 	versao varchar(255) NOT NULL,
-	categoria_id numeric NOT NULL,
+	categoria_id int8 NOT NULL,
 	CONSTRAINT qualificacao_pk PRIMARY KEY (id),
 	CONSTRAINT qualificacao_fk FOREIGN KEY (categoria_id) REFERENCES rh.categoria(id)
 );
