+++ server/src/main/resources/db/V6__create_experiencia.sql
@@ -1,6 +1,6 @@
 CREATE TABLE rh.experiencia (
-	profissional_id numeric NOT NULL,
-	qualificacao_id numeric NOT NULL,
+	profissional_id int8 NOT NULL,
+	qualificacao_id int8 NOT NULL,
 	tempo varchar(255) NOT NULL,
 	CONSTRAINT experiencia_pk PRIMARY KEY (profissional_id,qualificacao_id),
 	CONSTRAINT experiencia_fk_qualificacao FOREIGN KEY (qualificacao_id) REFERENCES rh.qualificacao(id),
